fix cr comments
