+++ packages/zendesk-support-adapter/src/filters/workspace.ts
@@ -31,7 +31,7 @@ const filterCreator: FilterCreator = ({ config, client }) => ({
   preDeploy: async changes => {
     await applyforInstanceChangesOfType(
       changes,
-      WORKSPACE_TYPE_NAME,
+      [WORKSPACE_TYPE_NAME],
       (instance: InstanceElement) => {
         instance.value = {
           ...instance.value,
@@ -47,7 +47,7 @@ const filterCreator: FilterCreator = ({ config, client }) => ({
   onDeploy: async changes => {
     await applyforInstanceChangesOfType(
       changes,
-      WORKSPACE_TYPE_NAME,
+      [WORKSPACE_TYPE_NAME],
       (instance: InstanceElement) => {
         instance.value = _.omit(instance.value, ['macros'])
         return instance
